widget/tests/test_bug673301.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

<body  xmlns="http://www.w3.org/1999/xhtml">
<p id="display"></p>
<div id="content" style="display: none"/>
</body>

<script type="application/javascript">
function getLoadContext() {
  return window.docShell.QueryInterface(Ci.nsILoadContext);
}

var transferable = Cc['@mozilla.org/widget/transferable;1']
                             .createInstance(Ci.nsITransferable);
transferable.init(getLoadContext());

transferable.addDataFlavor("text/unicode");
transferable.setTransferData("text/unicode", document);

Services.clipboard.setData(transferable, null, Ci.nsIClipboard.kGlobalClipboard);

transferable.setTransferData("text/unicode", null);

SimpleTest.ok(true, "Didn't crash setting non-text data for text/unicode type");
</script>
</window>
